Routing-related classes moved away from public classes into ToDoRouting inner classes.
Previous public classes contained minimal logic and were not re-used within the app, so it's better to have them as inner classes of ToDoRouting to minimize coupling between different files and also for reducing the number of Java files.
