--- architecture-examples/gwt/src/com/todo/client/ToDoRoutingAll.java
@@ -1,13 +1,0 @@
-package com.todo.client;
-
-/**
- * A routing function that matches all todo items.
- */
-public class ToDoRoutingAll implements ToDoRoutingFunction {
-
-	@Override
-	public boolean matches(ToDoItem item) {
-		return true;
-	}
-
-}
